debug

+++ GLControl/Palette.h
@@ -48,7 +48,7 @@ namespace GL {
 
 	public:
 		void add(int value_, lib::iPoint3D color_);
-		lib::iPoint3D get(int value_);
+		lib::iPoint3D get(double value_);
 		void getMinMax(float& fMin_, float& fMax_);
 		unsigned getInterpolate();
 		bool fillPalette(unsigned nPaletteID_);
